+++ app/Http/Controllers/KegiatanController.php
@@ -17,7 +17,9 @@ class KegiatanController extends Controller
     public function destroy($id)
     {
         $data = Kegiatan::find($id);
-        unlink($data->gambar);
+        if (file_exists($data->gambar)) {
+            unlink($data->gambar);
+        }
         $hapus = $data->delete();
         if ($hapus) {
             return redirect('/kegiatan');
